+++ lib/src/tuple/tuple_5.dart
@@ -117,6 +117,12 @@ class Tuple5<T0, T1, T2, T3, T4> extends Tuple {
   }
 
   @override
+  R map<R>(
+          R Function(T0 first, T1 second, T2 third, T3 fourth, T4 fifth)
+              callback) =>
+      callback(first, second, third, fourth, fifth);
+
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       (other is Tuple5 &&