+++ lib/src/tuple/tuple_7.dart
@@ -157,6 +157,13 @@ class Tuple7<T0, T1, T2, T3, T4, T5, T6> extends Tuple {
   }
 
   @override
+  R map<R>(
+          R Function(T0 first, T1 second, T2 third, T3 fourth, T4 fifth,
+                  T5 sixth, T6 seventh)
+              callback) =>
+      callback(first, second, third, fourth, fifth, sixth, seventh);
+
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       (other is Tuple7 &&