+++ lib/src/tuple/tuple_8.dart
@@ -173,6 +173,13 @@ class Tuple8<T0, T1, T2, T3, T4, T5, T6, T7> extends Tuple {
   }
 
   @override
+  R map<R>(
+          R Function(T0 first, T1 second, T2 third, T3 fourth, T4 fifth,
+                  T5 sixth, T6 seventh, T7 eighth)
+              callback) =>
+      callback(first, second, third, fourth, fifth, sixth, seventh, eighth);
+
+  @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       (other is Tuple8 &&