+++ test/math_test.dart
@@ -1,6 +1,7 @@
 import 'dart:math' as math;
 
 import 'package:more/math.dart';
+import 'package:more/number.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -204,6 +205,33 @@ void main() {
       for (var i = 0; i < max; i++) {
         expect(BigInt.from(i).isProbablyPrime, primes.contains(i));
       }
+    });
+    test('Complex', () {
+      const a002145 = [3, 7, 11, 19, 23, 31, 43, 47, 59, 67, 71, 79, 83, 103];
+      for (var i = 0; i < a002145.last; i++) {
+        final values = [Complex(-i), Complex(i), Complex(0, -i), Complex(0, i)]
+            .map((value) => value.isProbablyGaussianPrime);
+        expect(values, everyElement(a002145.contains(i)));
+      }
+      const norm5 = [
+        Complex(1, 2),
+        Complex(-1, 2),
+        Complex(1, -2),
+        Complex(-1, -2),
+        Complex(2, 1),
+        Complex(-2, 1),
+        Complex(2, -1),
+        Complex(-2, -1),
+      ];
+      for (var a = -5; a <= 5; a++) {
+        for (var b = -5; b <= 5; b++) {
+          final value = Complex(a, b);
+          if (value.norm() == 5) {
+            expect(value.isProbablyGaussianPrime, norm5.contains(value));
+          }
+        }
+      }
+      expect(const Complex(math.pi, math.e).isProbablyGaussianPrime, isFalse);
     });
   });
   group('lcm', () {